Added ShardedModel, BigIntegerField, and ShardedManager
A shardable model must inherit ShardedModel and have a BigIntegerField
with sharding_root=True. Saving this model will automatically add
a 'shard_key' hint with the value of the BigIntgerField and add a
_shard attribute on the instance itself. This is for writes.

For reads, a user must pass an instance into one of the following:
all(), get(), or filter(). This does something similar to the above
and adds a 'shard_key' hint.